Trade endpoint: add & edit (#3)

+++ pkg/entity/coin.go
@@ -3,5 +3,5 @@ package entity
 type Coin struct {
 	ID     string             `json:"id"`
 	Symbol string             `json:"symbol"`
-	Prices map[string]float32 `json:"prices"`
+	Prices map[string]float64 `json:"prices"`
 }
